requirements.txt
flask
python-dotenv
tavily-python
requests
fastapi
uvicorn
python-multipart
apscheduler
google-generativeai==0.8.6
groq